Process flattened trait properties like regular trait properties
Summary:
We were assuming that flattened traits properties could be added just
like regular properties, but then we noticed that re-declaring a
protected property via a regular property restricts its visibility to
subclasses of the redeclaring class, while doing the same via a trait
property does not. This resulted in fatals in repo-auth mode that
didn't happen in sandbox mode (fixed by D9328606).

But there are also additional consistency checks that apply to trait
properties, but not regular properties - so the new test case would
fatal in sandbox mode, but not in repo-auth mode.

This diff processes all the flattened trait properties exactly the
same way as regular trait properties, which will hopefully fix all the
inconsistencies.

Reviewed By: ricklavoie

Differential Revision: D9347762

fbshipit-source-id: 39e6dcbb8af62ccd591ea213e461c2cc262ab41f