Include address literal contents in IR dumps
Summary:
It's been bugging me for a while that these aren't here. We mostly use
a PC-relative LDR to materialize an address, and these values were missing
from the IR dumps.

This adds the literal value when the load is PC-relative and either a word or
double-word (though TC addresses should always be 32-bits wide)

I'm putting this functionality behind a boolean flag on vixl::Disassembler
because vixl includes fuzz tests that segfault when this is enabled. (and by
design, its disassembler is supposed to be safe even on invalid code, which we
don't leverage for IR dumps)

Since IR dumps are only used by developers and if the address isn't
dereferencable, the translation will crash anyways, I think it's Safe Enough™
but I could be totally insane.

Reviewed By: ottoni

Differential Revision: D9370786

fbshipit-source-id: 60e5a96ca86649eed6b43e1baa37dfb7b0b9c2ba