Fix binary expression argument indentation
Summary:
In D7527211, we attempted to indent continuations of multiline binary expressions when passed as arguments, in order to help distinguish arguments from one another.

I used this as an example in the summary:

```
// before D7527211
fn(
  $foo === 1 &&
  $bar === 2,
  $baz === 3 &&
  $qux === 4,
);

// after D7527211
fn(
  $foo === 1 &&
    $bar === 2,
  $baz === 3 &&
    $qux === 4,
);
```

But I should have added a test case verifying that this was the result we produced, because it isn't. After D7527211, we actually produce this:

```
fn(
  $foo === 1 &&
  $bar === 2,
  $baz === 3 &&
    $qux === 4,
);
```

This happens because I neglected to actually use the new argument to `handle_list`, `handle_element`, when transforming any item except the last.

This change ensures that we actually invoke `handle_element` on each item.

Reviewed By: arxanas

Differential Revision: D9356473

fbshipit-source-id: 4fdda01d0530bfad9ca8fb0b2713c564fc3f60a7