Add splits within braces less often for single braced items
Summary:
In D7527281, we permitted preserving a no-surrounding-whitespace style for braced items like subscript indexing, but only when the subscript expression is a variable or literal. I think we could extend this behavior to any node which is not blacklisted by the `looks_bad_in_non_parental_braces` function used by D9231458. This allows us to preserve the no-surrounding-whitespace style in array and vec type specifiers (a style which, in D9231458, we already started preserving in type specifiers for user-defined classes and builtins with multiple type arguments like `dict`).

Additionally, it seems that D7527281 overlooked adding this support to `transform_braced_item`'s cousin `transform_braced_item_with_trailer`, so this change is done in both.

Reviewed By: arxanas

Differential Revision: D9231819

fbshipit-source-id: a850d44fddc6e79beb3a4dfd2988b1e77e9d81c5