Ensure that spaces in leading XHP trivia are preserved
Summary:
Semantically-significant whitespace can occur in leading trivia for XHPBody tokens when that whitespace follows an XHPOpen node or a braced expression. We're currently inserting a `Split` at these locations, but this introduces a semantic change when we choose not to split, since we will not add any whitespace.

Instead, we should do the same as we do with trailing trivia for XHPBody tokens--preserve newlines and spaces when present.

Reviewed By: arxanas

Differential Revision: D9380842

fbshipit-source-id: ca01e9c4d1b70b17c1f029e4f3465ea6c62c7594