Add bytecode emission and resolution for reified generics
Summary:
This diff is the first stage to emitting bytecode for reified generics.

```
function f<reified T>() {
  <body>
}
function g<reified T1, reified T2>() {
  f<reified (T1, T2, int)>();
}

g<reified int, reified string>();
```
This diff introduces `CombineAndResolveTypeStruct` instruction which enables merging and resolution of type structures that is needed for the above example. In function `g`, we need a way to merge two reified generics, namely `T1` and `T2`, into the tuple with 2 holes in it as well as an int.

`CombineAndResolveTypeStruct` instruction requires holes on the type structure to be appropriately IDed, in this case T1 would be id 1 and T2 would be id 2 which is denoted by stack push order. And a constant type structure to be pushed to the stack with these IDs as holes in it.

Crudely, it would look something like:
```
CGetL T1
CGetL T2
Array 1
CombineAndResolveTypeStruct 3
```

this series of bytecodes will turn into a single type strcuture on the stack.

The following diffs continue on this by adding the same mechanism for classes and actually using the reified generics in the function body.

Reviewed By: paulbiss

Differential Revision: D8941709

fbshipit-source-id: 960631f152bc359f9ab9d7d79b3c63e98b45fe3a