Default FFP to true in hh_single_type_check
Summary:
This diff is the start of deleting the legacy parser. We make the default of hh_single_type_check to FFP.

There are a few test changes, but all of them are things we have tasks open about. I'd prefer that these just get moved and changed after the tasks update them rather than having to keep track of ffp vs. legacy any longer.

Reviewed By: pittsw

Differential Revision: D9403477

fbshipit-source-id: db0af7bc012f281dfa71be8879ac3148e581fb73