Fix naming of $$ in ternary operator
Summary:
D8028383 introduced a regression in naming of `$$` inside the ternary operator.

The pipe operator requires a `local_id` for `$$` in the `Pipe` AAST node, this `local_id` was incorrectly previously, leading to `$$` being untyped in the typing phase.
This was due to the incorrect use of `Env.found_dollardollar` function. This function generates a new `local_id` for the `$$` variable. What is intended is that we inject the existing `local_id` for ``$$` into the environment, which is what this diff is trying to fix.

Consider the following code:
```
  1 |> (true ? $$ : $$)
```
Previously, there were 3 `local_id`s for `$$`, one for each branch in `Dollardollar` AAST node and an extra one in `Pipe` AAST node (introduced when merging the two branches). So the `local_id` inside `Pipe` node is typed, but the ones in `Dollardollar` nodes don't pick up the type and fall back to `Tany`.
Now, there is only one `local_id` for `Dollardollar` and `Pipe` AAST node. So the typing phase will be able to pick up existing typing and not use `Tany`.

Interestingly the following code works:
```
  true |> ($$ ? $$ : $$)
```
because the `$$` is correctly named and `$$` is present in the environment after typing the condition. The branches are able to pick up the correct `local_id` and hence the typing.

Reviewed By: nbenton

Differential Revision: D9420725

fbshipit-source-id: 1abe210101fb28715dbc9273bbf629f4f2b7c9b3