Forbid many expressions from being interpolated in strings v2
Summary:
The set of things you can interpolate in PHP strings is pretty arbitrary, and varies depending on which mode of interpolation you used.

A lot of the rules have been reversed engineered. The official PHP specification doesn't actually match up with reality. There's not an obvious set of rules dictating which things can be interpolated and when.

This is a resubmission of {D8986090}, which was reverted because some HHVM tests silently failed. I've incorporated the failing tests' contents into the test cases for this diff, which I'll mark below in comments.

Reviewed By: pittsw

Differential Revision: D9411797

fbshipit-source-id: 6acc69b46804c8fe8882fae9f8e0435ef9a730cf