Add SetRangeM final member operation
Summary:
SetRangeM is used to copy raw binary data from a variety of sources
into a string: bool, int (1, 2, 4, or 8 bytes), double (4 or 8 bytes),
vec<bool>, vec<int>, vec<double>, or another string. For vecs and strings, the
data can optionally be reversed as it's being copied. Data is stored using the
machine's native byte ordering. For a detailed description of the exact
semantics, look at `SetRangeM` in `bytecode.specification`.

Currently, all of the work is done by the SetRange function in
member-operations.cpp; it should be fairly straightforward to add JIT support
for any cases that become particularly hot.

Reviewed By: jjgriego

Differential Revision: D9371858

fbshipit-source-id: 7e5e0ff4a7327beb50114376f605b94f11beaf86