Reject static properties that use generic parameters in their types
Summary:
Hack should reject any static property whose type contains a generic parameter from the enclosing class.
Unfortunately, it did not do so for any "constructed" type e.g. an instantiation of a generic type such as `C<T>` or a built-in parameterized type such as `vec<T>`.

Moved check to typing phase so that we can deal with properties "inherited" through `use` on traits. Also ignore presence of type parameters in property types in traits themselves, as these are benign e.g. we can instantiate the trait type parameter with a ground type in the `use` clause.

Reviewed By: nbenton

Differential Revision: D8509883

fbshipit-source-id: 292efd3b4278203b9a396f978005552da41e105a