HHVM Debugger: support relative breakpoint paths
Summary:
This adds experimental support for breakpoints that contain only relative or filename paths. The VS Code Debug Adapter Protocol does not currently support such breakpoints, but hphpd does and we'd like to be able to offer feature parity.

A breakpoint in "foo.php" will attempt to set in any compilation unit whose filename is "foo.php", even if the rest of the paths differ. A single breakpoint can resolve to multiple locations as units are loaded, but only one verified event is sent to the debugger client because (currently) the protocol only allows for a single breakpoint to resolve to a single location.

We'll update this with a protocol enhancement to allow the debugger backend to communicate multiple installed locations for a single breakpoint in a future change.

Reviewed By: velocityboy

Differential Revision: D9473646

fbshipit-source-id: 3709085b8d1759ef75bc5eaadcd152cf1e5bd4a0