Rework smashable instructions on ARM to enable smashing code that is executing
Summary:
Reading the ARMv8 architecture manual, I found that it's safe to patch
instructions that are already executing in some special cases.  This diff
changes the smashable code sequences on ARM to take advantage of these special
cases.  More specifically, smashable jumps, branches and calls can now be
patched to their direct versions (when the offset of the target fits in the
immediate), which avoid the LDR + BR/BLR patterns previously used (thus saving a
load and an indirect call/jump).

The new scheme requires inserting veneers (basically a LDR + BR sequence), which
are emitted at the end of each translation (similar to literal pools), and the
smashable jump/branch/call instructions are emitted as direct ones and initially
go to a veneer.  When patching the code to the correct target, the literal
loaded by the veneer is patched and, if the offset to the target fits in the
direct jump/branch/call immediates, these instructions are patched to go
directly to their targets.

With this change, the patterns used by far jumps and branches don't match the
smashable sequences anymore.  Far jumps and branches are still emitted as
before, and potentially optimized by the relocator, but they're now called "far
jumps" instead of "smashable-looking jumps".

Another consequence of this change is that the optimizations that were performed
by the optimizeSmashed{Jmp,Jcc,Call} functions when pre-smashing optimized
translations in retranslate-all mode don't apply and aren't needed anymore.
There's one exception for conditional branches, which can still be optimized
further in some cases (1MB < offset < 128MB).

Reviewed By: jjgriego

Differential Revision: D9393913

fbshipit-source-id: 53e6a33087d7272c103ec66ac694522d385031f5