Clean up trait substitution
Summary:
A lot of code is shared between hhbbc and the runtime via a template
helper class. That class was originally written to share code between
hphpc and the runtime, and is far more general than is currently needed.

This detemplatizes and simplifies a lot of the code involved.

The only intended change in behavior is that there were two
errorUnknownMethod functions with very slightly different error
messages - `Unknown trait method` vs `unknown method`. I'm pretty sure
this wasn't intentional.

Reviewed By: ricklavoie, alexeyt

Differential Revision: D9498086

fbshipit-source-id: 3f175d4e4395e7192a31938760abe80e0ffaaf62