Fix simplifier assertion
Summary: This should be safe, if the output Tmp from CreateSSWH is available then so is its input. I think a cleaner fix here would be to move this logic to load-elim and add a memory-effect for creating a SSWH, it would probably expose more opportunities for this to kick in too.

Reviewed By: ricklavoie

Differential Revision: D9512595

fbshipit-source-id: 26e223beaadc1a7454bf39f7797b16e44472f233