Option to trace all parsed files
Summary:
I tried debugging some recent issues [1][2], and didn't get anywhere, because
while Mercurial logs have enough info to (I believe) reproduce what they were
doing  (what was the revision they were based on, what was the stack of local
commits, what revision they rebased to), repeating the same steps doesn't
reproduce the error.

We need more logging to gain insight what did hh_server actually do when things
happened. Ideally, we would like to see what were the contents of each parsed
file when it was parsed, but that will obviously be too much. I'll aim for a
middle ground, and log names of all top level entities in parsed files.

It will make logs much bigger (hence putting it behind config flag, just in
case), but

a) the amount of stuff logged is still O(change), with no fan-outs, so we
should never attempt to log all the names in codebase
b) this is roughly similar level of logging that Mercurial/Nuclide already do,
so we should be fine too

The one thing that we might miss is that this doesn't log lazy parsing (it can
fan-out more than O(change)), but let's worry about it later.

Reviewed By: vassilmladenov

Differential Revision: D9517020

fbshipit-source-id: d2c31c53e3a7c634a34d94a6eb94e80d1fe6d6f0