Add sampleModel to Infer class and rewrite it
Summary: We currently cannot `sample` from a `Model` class because `Model` is special and because HackPPL does not have models implement the `Distribution` interface. As a result, in order to allow for cross model sampling, we need to add a special function to do so. Currently the proposal is for a `sampleModel` or its equivalent.

Reviewed By: michaeltingley

Differential Revision: D9000243

fbshipit-source-id: f21296d7af3513fcf743ccf1560c52195672a881