Eliminate circular dependency in emit_call_expr
Summary:
There is a potential circular dependency between emit_call_expr,
emit_flavored_expr and emit_expr involving the same `expr` node. This does not
happen in practice because emit_call_expr is used only for expressions not
matching is_special_function, so emit_flavored_expr is guaranteed to not use
the generic emit_expr code path. But it is fragile, as the A.Call rules of
emit_expr need to match the logic of is_special_function.

Let's improve this by reversing the relationship of emit_call_expr and
emit_flavored_expr. The responsibility of boxing/unboxing was lifted from
emit_call_expr to emit_expr, the remainder was just a call to
emit_flavored_expr, which got inlined, allowing the relationship reversal.

I am keeping emit_call_expr even as it was reduced to a trivial call to
emit_call, as the next diff will move all A.Call rules from emit_expr to
emit_call_expr, allowing removal of is_special_function logic.

There's a minor change in emitting position for box/unbox. We now use the same
position as used by emit_call instead of using the last_pos override. This
seems to be better than before and more consistent with other code. Anyway, it
won't change anything in practice.

This also allows addition to optional async eager label to emit_call without
having to add it to emit_expr.

Reviewed By: oulgen

Differential Revision: D9519567

fbshipit-source-id: ace3f0b7b8f37dc59d0902f5717e89aba49a4e56