Fix array cast and array_key_exists on ArrayIterator, add tests.
Summary:
We didn't behave the same way as PHP when casting an ArrayIterator with the STD_PROP_LIST flag set to array (it should have behaved like ArrayObject bit didn't) https://3v4l.org/plurC
Because we depend on that in our array_key_exists implementation, we had a related bug there https://3v4l.org/GBmXK
Fix both problems by unifying the codepaths; add tests.
If you managed to construct a recursive ArrayObject or ArrayIterator, we would run out of stack and segfault when you tried to cast it to array; add a recursion check so that we throw instead; add a test.

Reviewed By: ricklavoie

Differential Revision: D9522808

fbshipit-source-id: 55ee61e569e2f07856b6986f2c62bfd4c229158c