fix assignment in conditionals + refinements with conjunctions/disjunctions of conditionals
Summary:
Assignements in conditionals, like
```
if ($a = $b) {}
```
are simply not accounted for at all. The assigned variable does not get added in the locals, and therefore a call to get_local on it simply returns TAny, silencing many possible errors.
This diff fixes that.

This diff also fixes refinements with conjunctions (&&) and disjunctions (||) of conditions: at least the first condition of the con/disjunction should refine.

Reviewed By: Wilfred

Differential Revision: D7774523

fbshipit-source-id: 1edba3eefb7f2983439acbad2dd1ac8d95c5e54d