make hack case sensitive for keywords
Summary:
If backend is not the codegen, enforce that all keywords are lowercase.

If the lexer scans a keyword, we check that the source text of the keyword is all lowercase.

Perf impact:
* Test suite runs in 134 s instead of 126 s
* Total initialization time to typecheck www: 468 s instead of 448 s

These perf regressions are only slightly higher than the one we would get in V1 of this diff, where we would simply not recognize capitalized keywords as keywords:
* Test suite : 132 s instead of 126 s
* Total initialization time to typecheck www: 461 s instead of 448 s

Compared to V1, this version also finds more errors. For example:
```
CONST MYCONST = 1;
```
`CONST` would not be parsed as the `const` keyword in V1, but maybe as a type name?
```
$x = (String) $y;
```
`String` was not recognized as the built-in `string` type, but as a user-defined type?

Reviewed By: jamesjwu

Differential Revision: D9347620

fbshipit-source-id: 6d4c43acce016985b5655c433056d2520217a6e6