Add hackfmt support for single line if statements (#8277)
Summary:
Currently, a single line if statement will be split into 2 lines after hackfmt is run. For example,
```
if (true) print("Hello, world!");
```
becomes
```
if (true)
  print("Hello, world!");
```

This PR changes hackfmt so it keeps single line if statements, as long as the line isn't too long. If the line exceeds the specified limit, then it'll be split into 2 lines.

This was tested on a file with both single line if statements (of various lengths) and multi-line if statements.
Pull Request resolved: https://github.com/facebook/hhvm/pull/8277

Reviewed By: fredemmott

Differential Revision: D9218213

Pulled By: pittsw

fbshipit-source-id: 24ed648924ab75735f4ec6e911bf157d28f737c6