Use local-only disk and parser heaps for typechecking.
Summary: During the typechecking phase (especially in lazy inits) we end up parsing and decling a bunch of files on the fly.  Many of these files will only be used a handful of times, so it's not really worth it to cache their on-disk representation and their parse tree.  Once we have the decl, it'll get put into the class heap as normal, and execution will continue at its full speed.  In exchange we get to avoid the overhead of writing to shared memory and the additional memory usage it needs to store all this stuff we don't really need.

Reviewed By: 2BitSalute, dabek

Differential Revision: D9548128

fbshipit-source-id: 0a88434f79a2211411b770ce30e322392f218f05