Clean up unit creation
Summary:
While making UnitEmitter::create const I noticed that it had a member
m_fMap which was used to create a map between FuncEmitters and emitted
Funcs, which it then uses in conjunction with a vector of all
FuncEmitters to build the Unit's m_funcTable.

It seemed better to plumb a local map through to UnitEmitter::newFunc
and get rid of the (now) mutable member. But it turns out that the
only use of m_feTab was to iterate all the FuncEmitters and add the
corresponding Funcs to m_funcTable. So get rid of both m_feTab and
m_fMap and just append the Funcs to Unit::m_funcTable as they're
created.

Reviewed By: paulbiss, alexeyt

Differential Revision: D9583092

fbshipit-source-id: b2045591a8ba6e0df2470def07bb5fb71442e95f