Prechecked files: track which files were checked at least once to prevent duplicate work
Summary:
If prechecked files infrastructure causes a file to be added to `needs_recheck`, it's because it "should have been rechecked at the init, but we skipped it". But if we DID recheck this file at any other point between init and now, there is no need to re-add it again now.

(Note: this ignores all push blocking failures!)

Reviewed By: pittsw

Differential Revision: D9322210

fbshipit-source-id: 37166042b10ecf9bc9ab20054b952c4a73ed923f