Remove phijcc
Summary:
phijcc is a phijmp which can conditionally branch to two different
phidefs. This is not how phis are usually treated and complicates analysis. It
is only used in a single place (constructing a do-while loop), and can easily be
replaced by a jcc to two blocks which each do a phijmp to their
destination. Indeed this is often how the final code ends up, because often one
needs to split the blocks anyways. Remove phijcc and replace its single usage
with manually split blocks.

Reviewed By: mxw

Differential Revision: D9628465

fbshipit-source-id: 3cc19c9a3b41009daf1b225a3767e06e0287a256