Move AttrHasForeignThis from Attr to a bit on Func.
Summary:
AttrHasForeignThis isn't being used by the emitter or by hhbbc, so we can move it to a bit on Func (not Func::SharedData because the value is different on different clones).
In order to keep closure cloning working correctly, define enum class Class::CloneAttr, which replaces the direct use of Attr in the closure cloning logic. It lets us distinguish between the closure clones created by CreateCl vs Closure::bindto, and in the latter case between clones that are static and/or have a foreign $this.

Reviewed By: ricklavoie

Differential Revision: D9625193

fbshipit-source-id: 8fe9430b7ce417bb0cf7e2921585143c9440d4ac