Improve typing of array appends
Summary:
Array appends are somewhat broken and confusing at the moment.  For example, you cannot append to a collection of an unresolved type (see `array_append_unresolved_good.php` below).  Also, because PHP and Hack arrays enjoy the copy-on-write semantics, it should be possible to append a value of any type `t` to say a vec of elements of type `tv`, and this should produce a vec of elements of type  `tv | t` (provided the vec is not a property of a class and has its type constrained by the property declaration).

This diff attempts to improve on these issues (and simultaneously to bring the implementation in agreement with the formal spec; see D8723456).

Reviewed By: CatherineGasnier

Differential Revision: D9082246

fbshipit-source-id: 43b810e1c36efbf486d2562be00e761763bf84a1