Add small language of subtype propositions to checker
Summary:
Generalize the result of `simplify_subtype` to a small language of propositions about types, supporting conjunction, disjunction, subtype and equality assertions, and `Unsat f`: essentially `false` but with an error message function attached.

Disjunction not yet used or implemented. Next diff will make use of this in place of all the `try_error` code in subtyping.

Reviewed By: manzyuk

Differential Revision: D9630559

fbshipit-source-id: 0f7594ff66ba662f1fad05a3ea19e187da941988