avoid calling `expr` twice on same expression
Summary:
Obviously, calling `expr` twice on the same expression is wrong because `expr` may change the type of locals, for example if the expression being checked is an assignment.

However, the `condition` function freely calls `expr`, on expressions which may have been checked elsewhere, sometimes resulting in duplicate calls to `expr` on the same expression.

See the added test cases for examples of where this may be problematic.

To avoid having to scrutinize how many times `expr` is called and to enforce that `condition` does not call `expr`, the solution we implement here is to pass a typed expression (i.e. a tast node) to `conditon` instead of an expression (i.e. a nast node).

Reviewed By: andrewjkennedy

Differential Revision: D8223160

fbshipit-source-id: 36fb6fbca1b445ba7a8c88c8bedc2d8d7a6606f7