fix watchman event watcher tests by adding "hgevents" to .hgrc
Summary:
Watchman event watcher is an additional service that Hack server can query to learn whether we are in the middle of source control update.

When testing D9632724 I noticed that it's tests are failing. I used guide[1] to verify whether it's also broken in prod - fortunately no. I tried it on a month-old revision, and it was also broken, which made me think that something changed in Mercurial or Watchman configuration. I looked at .hgrc of www and picked a line that seemed relevant - copy pasting it here fixed the tests.

Reviewed By: CatherineGasnier

Differential Revision: D9632875

fbshipit-source-id: 73cd4b4c56c9661e98901513685d1525e29eba7b