Test case showing that race conditions with disk can corrupt server state and everything is broken
Summary:
In comments of [D9548128](https://our.intern.facebook.com/intern/diff/D9548128/?transaction_id=2052069021511397), I was trying to explain what `File_heap` is for. My general impression was that it "helps prevent race conditions when disk contents change under our feet, but it is only a "best-effort" thing, and race conditions sometimes happen. We sometimes notice them and crash, and hope that there are no cases where things can silently go wrong".

This explanation didn't feel very satisfying, so I tried to think really hard and formalize what is actually happening, and I'm happy (?) to report that things indeed can silently go wrong, and one can never be sure about any invariants and consistency of server state. Great!

Here is a bit of historical context:

Long, long time ago, we used to read every single file version exactly once - during init, we would parse every file, compute and store their ASTs, and everything would refer to this AST when necessary. The AST in memory would be updated when file changes.

This was great for correctness - since there was single source of truth that we fully control (ASTs in shared memory), we were guaranteed that all server state (parse trees, naming table, declarations) was consistent.

With growth of www, we had to do away both with parsing all files during init (via saved state), as well as with storing all ASTs (via lazy parsing). The source of truth is now out of our control, and this can lead to nasty race conditions, the most common one being:

- naming table claims that definition of X is in file X.php
- in the middle of some computation, we need X so we try to declare X, but when looking in X.php it's no longer there (because disk changed it in the meantime)
- now what?

Ideally, when noticing such situation, we should roll-back the computation, process the change, and restart it. But many (most) operations in the server are of destructive nature, and can't be easily rolled back, or safely interrupted.

The next best thing is to just crash, which we actually do, tens of times per day (see Decl_not_found and File_heap_stale exceptions). It happens (or that's what we tell ourselves) mostly around rebases, since this is when we are both being busy doing things, and a lot of files keep changing. We crash only in cases where a declaration we expect is missing. If the discrepancy is of other nature (i.e return type of function is different during two accesses), we just keep going and hope that it doesn't corrupt the state in any way. The test case in this diff shows that it can, in fact, silently corrupt the state and lead to missing errors.

So going back to original question, what does FileHeap do? It tries to mitigate the problem a bit, by caching file contents (so once again, we have a single point of entry that we control), and two successive reads will always yield consistent parse trees. But since FileHeap starts empty, there is no guarantee that the first read of the file will be consistent with data from saved state. So it can prevent a subset of race conditions, but not all of them.I have vague recollection that one such race condition might have been  triggered pretty regularly by code generators / build scripts that were working by deleting, and almost immediately re-creating some files. This is the "two reads in a row" scenario that FileHeap might be able to save us from.

So what we should do about it?

- Right now, FileHeap stores (FOREVER) every file that we've ever happened to look at. I highly suspect that there is a limited upside to it. If the only case that is actually helped by it are code generators, then maybe storing just contents of files that changed (in Parsing_service) would be enough? Digging through old diffs to find why did we put this things there in the first place might be helpful... similarly, keeping those things forever looks like overkill, especially if it doesn't really solve the underlying problem. We should be able to evict them.

- It fills me with great discomfort that you can't really say anything about the consistency and correctness of the system because of race conditions, and it makes many problems undebuggable. I think it would be very valuable to **restore some sanity by always maintaining a hash that we expect file to have** (we can afford to deliver all hashes with saved state), and any any time we have a mismatch log it (so that any subsequent bugs can be attributed to it). If this doesn't happen too often, we can just start crashing on any such mismatches similarly to how we already crash on missing declarations. At some distant point in the future, we can also consider identifying cases where we are actually able to safely recover from those mismatches.

(Note: this ignores all push blocking failures!)

Reviewed By: arxanas

Differential Revision: D9601566

fbshipit-source-id: 831b3c63bb4799632bede7d2748383c323bb83f4