Initial Support for Zstd Compressed Content-Encoding
Summary:
The zstd format and Content-Encoding ("`zstd`") are being standardized in
https://tools.ietf.org/html/draft-kucherawy-dispatch-zstd-03. The draft has
been accepted and will be published shortly.

This diff adds initial support to HHVM for this content encoding, which will
allow for further experimentation and development.

For the purposes of experimentation, this diff will cause HHVM to prefer
responding with zstd-compressed data when clients include both `zstd` and `br`
in the request's `Accept-Encoding` header. This will only occur when zstd
compression is enabled, which is not the case by default.

Cribbed heavily from D3010183.

Differential Revision: D9386480

fbshipit-source-id: 8c35cdf5a97353b67290f2cdc66709a492f7aa6c