handle possible absence of Next continuation using exceptions
Summary:
Local lookup happen in the `Next` continuation. If there is no such continuation (that happens when the code is actually dead), so far we throw a `Continuation_not_found` exception.

So far, we arrange to never hit that case and never throw this exception, using functions `if_next_expr` and `if_next_condition`. However, this becomes easily intractable as we want to detect more dead code (for example in D8330906 further up this stack of diff). Besides, the risk of throwing this exception in production is too big.

Instead, in this diff, we get rid of `if_next_expr` and `if_next_condition` and simply catch the exception in `expr`. If we catch this expression, we return type `Terr` for the expression. Catching the exception at the expression level allows to still typecheck the rest of the statement. It is not necessary to catch the exception at the statement level as statements never perform local lookups directly.

Reviewed By: kmeht

Differential Revision: D9607957

fbshipit-source-id: f36fab4c8c3eb1eacb44d284f1c6119e379454a2