Compacting sharedmem GC
Summary:
The current sharedmem GC requires a significant amount of additional memory, as
it copies all live values to a new temporary heap.

In low-memory scenarios, this is problematic as the OS may have trouble
providing the necessary space. Furthermore, the majority of sharedmem usually
survives GC, as it is more like an old-gen than a young-gen.

This diff replaces the existing copying collector with a compacting one, which
has the same linear time complexity as the copying collector and requires no
additional space by doing a bit of pointer juggling.

As the mutator runs, the sharedmem segment will become fragmented as entries
are deleted over time. The compacting collector eliminates this fragmentation
by moving values "to the left" until there are no gaps.

Moving entries around poses an issue: the hash table contains pointers to these
entries which become invalid if the entries move. Thus, when we move entries,
we also need to fix up the pointers in the hash table.

The compacting GC first marks all the live entries. Each heap entry is prefixed
with a 8-byte header. The header's low-order bit always set. During marking,
this header is copied into the address member of the hash table entry (which
will be invalid) and is replaced with a pointer to the address member which now
contains the header.

Because the address member is aligned, its address's low-order bit will always
be unset. Thus, we can differentiate between marked heap entries and unmarked
ones by inspecting the low-order bit of the heap entry's 8-byte prefix.

After marking, we walk the heap, stepping over unmarked entries (lsb 1) and
moving marked entries (lsb 0) to the left. Before we move, we chase the pointer
back to the original address member, copy the header back and store the new
address.

Reviewed By: gabelevi, pittsw

Differential Revision: D9628787

fbshipit-source-id: 6a005f5e7ebc61bc8d438526ce5a2f98d7960971