Refactor XHP body handling
Summary: The handling for XHP bodies is somewhat convoluted for two reasons: one is that the parser changes the specification for how trivia is lexed within XHP bodies, and the second is that the parser historically did not conform to its own specification. Recently, the parser has been fixed and now behaves consistently, and the formatter no longer needs to work around the incorrectness of the parser. It needs only to deal with the awkwardness of managing semantically-signficant whitespace and the slightly-modified meaning of leading trivia in XHP bodies. This is a good opportunity to clean up this mess!

Reviewed By: pittsw

Differential Revision: D9380982

fbshipit-source-id: 0c72a00b5375fb8205c17e92fcae51c9d8630daa