fix infinite loops
Summary:
`Nast_terminality` can recognize infinite loops to deal with terminality. In order to kill this module, we need to recognize infinite loops as well in the way we deal with control flow.

The cleanest and simplest way to do that is to add cases in the `condition` function (dealing with refinements): if the refinement is always false, simply discard the `Next` continuation.

One side effect of that is that the typechecker will also become clever about code like
```
if (true) {
  ...
} else {
  ...
}
```
but since we need to be clever about code like
```
while (true) {
  ...
}
```
it seems reasonable to be clever about the former case with `if` as well.

Reviewed By: andrewjkennedy, Wilfred

Differential Revision: D8330906

fbshipit-source-id: 5af59cb476c83f48f4e58dd26faa9f7a930bd7e8